Update to ares v111r07 release.
  - ruby: removed Input::onHotplug(); luna and lucia handle hot-plugging
    automatically
  - Famicom: refactored SxROM and MMC1 into a single class
  - Famicom: converted generic class variable types (bool, uint) to
    uint# types
  - Game Boy Advance: fixed a bug with keypad interrupt handling (keypad
    array was erroneously marked static)
  - all: moved inline designators from function designators to function
    declarations

+++ ares/fc/ppu/render.cpp
@@ -60,7 +60,7 @@ auto PPU::renderSprite() -> void {
   if(!enable()) return;
 
   uint n = latch.oamIterator++;
-  int ly = io.ly == vlines() - 1 ? -1 : io.ly;
+  int ly = io.ly == vlines() - 1 ? -1 : (int)io.ly;
   uint y = ly - oam[n * 4 + 0];
 
   if(y >= io.spriteHeight) return;
